+++ CHANGELOG.md
@@ -14,9 +14,20 @@ All notable changes to this project will be documented in this file.
 * Updated all Bible book names in the Italian New Diodati translation to match canonical formatting.
 
 ---
-## [4.0.3] - 2025-08-19
+## [4.0.4] - 2025-08-19
 ### Fixes
 * Removed unnecessary debug statements.
 * Resolved an issue where song lyrics were not rendering correctly on windows systems.
  
  ---
+
+ ## [4.0.4] - 2025-08-17
+### Added
+* Added countdown timer to song line duration.
+
+### Fixed
+* song lyrics search improved.
+
+---
+
+